src/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
boto3==1.17.14
botocore==1.20.14
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
dill==0.3.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
html-testRunner==1.2.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
multidict==5.0.0
multiprocess==0.70.10
multithread==1.0.1
mypy==0.800
mypy-extensions==0.4.3
nester==1.0.0
nltk==3.5
numpy==1.19.4
pandas==1.2.0
pexpect==4.8.0
Pillow==8.1.0
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
snowballstemmer==2.1.0
soupsieve==2.1
threadpoolctl==2.1.0
tqdm==4.55.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
yarl==1.6.2
